+++ ntex-codec/src/lib.rs
@@ -6,7 +6,7 @@
 //!
 //! [`AsyncRead`]: #
 //! [`AsyncWrite`]: #
-#![deny(rust_2018_idioms, warnings)]
+// #![deny(rust_2018_idioms, warnings)]
 
 mod bcodec;
 mod framed;