+++ ntex/CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## [0.1.5] - 2020-04-07
+
+* ntex::http: enable client disconnect timeout by default
+
+* ntex::http: properly close h1 connection
+
+* ntex::framed: add connection disconnect timeout to framed service
+
 ## [0.1.4] - 2020-04-06
 
 * Remove unneeded RefCell from client connector